refactor(sms_count): shifts default msgsrvcid to be from config, helper method defined in test
