+++ src/app/services/sms/sms_count.server.model.ts
@@ -2,6 +2,7 @@ import { parsePhoneNumberFromString } from 'libphonenumber-js/mobile'
 import { Mongoose, Schema } from 'mongoose'
 import validator from 'validator'
 
+import { smsConfig } from '../../../app/config/features/sms.config'
 import { FORM_SCHEMA_ID } from '../../models/form.server.model'
 import { USER_SCHEMA_ID } from '../../models/user.server.model'
 
@@ -42,7 +43,7 @@ const VerificationSmsCountSchema = new Schema<IVerificationSmsCountSchema>({
 VerificationSmsCountSchema.pre<IVerificationSmsCountSchema>(
   'save',
   function (next) {
-    const formTwilioId = process.env.TWILIO_MESSAGING_SERVICE_SID
+    const formTwilioId = smsConfig.twilioMsgSrvcSid
     this.isOnboardedAccount = !(this.msgSrvcSid === formTwilioId)
     return next()
   },
